games_logic_back/requirements.txt
uvicorn[standard]
pymongo
python-dotenv
pydantic
python-socketio[asgi]